legion-gui/pyproject.toml

# Legion — decentralized, anonymous communication platform
# Copyright (C) 2026  n3ur0-f4ll
#
# This program is free software: you can redistribute it and/or modify
# it under the terms of the GNU Affero General Public License as published
# by the Free Software Foundation, either version 3 of the License, or
# (at your option) any later version.
#
# This program is distributed in the hope that it will be useful,
# but WITHOUT ANY WARRANTY; without even the implied warranty of
# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
# GNU Affero General Public License for more details.
#
# You should have received a copy of the GNU Affero General Public License
# along with this program. If not, see <https://www.gnu.org/licenses/>.


[tool.poetry]
name = "legion-gui"
version = "0.1.0"
description = "Legion GUI — pywebview interface for the Legion node"
authors = []
license = "AGPL-3.0-or-later"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"
pywebview = "^5.0"

[tool.poetry.scripts]
legion-gui = "app.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py312"]
